feat: 数据库分表定义 SQL

+++ DspLib/DataBase/Model/SeedInfo.cs
@@ -2,7 +2,7 @@
 
 public class SeedInfo
 {
-    public int SeedInfoId { get; set; }
+    public ulong SeedInfoId { get; set; }
     public List<SeedGalaxyInfo>? SeedGalaxyInfos { get; set; }
     public SeedPlanetsTypeCountInfo? SeedPlanetsTypeCountInfo { get; set; }
     public SeedStarsTypeCountInfo? SeedStarsTypeCountInfo { get; set; }
